Add files via upload
This is a project my colleagues and I did for a university subject called E-Learning. The project is supposed to help you study for the Greek admission exams, for the subject of history. You are given the opportunity to study the subject matter and then take tests (multiple choice), different test for every chapter. Afterwards you can check the correct answers (which are shown in the statistics.txt file) and see which chapters you need to study more.
P.S. Many files (not the .JAVA) ,such as the manual , the handbook and the chapters you are supposed to study, are written in greek, because the project focuses on helping the Greek students.